Remove PIO initialization
The I2C PIO program initialization has been removed to better suite
situations where there are other devices on the I2C bus the display
uses. Init has to be done outside the library, and relevant configs
(PIO, state machine, and a pointer to the I2C write function have to
be passed as parameters.